一時的にツイート削除を無効化

+++ WebContent/inner/TDeleteContent.jsp
@@ -1,6 +1,10 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <script>
     function DeleteContent(nUserId, nContentId, bPreviousTweetExist) {
+	if(!window.confirm('<%=_TEX.T("IllustListV.CheckDelete")%>')) return;
+	DeleteContentBase(nUserId, nContentId);
+	return false;
+/*
         DeleteContentInteractive(
             nUserId, nContentId, bPreviousTweetExist,
             '<%=_TEX.T("IllustListV.CheckDelete")%>',
@@ -10,5 +14,6 @@
             '<%=_TEX.T("IllustListV.CheckDeleteTweet.Yes")%>',
             '<%=_TEX.T("IllustListV.CheckDeleteTweet.No")%>'
         );
+*/
     }
 </script>
